[feat/delay-logout] add constant for default end_dynasty
